fix: 회원가입과 fcmToken 등록이 동시에 실행되는 현상 해결
