+++ presentation/signup/src/main/java/com/depromeet/threedays/signup/SignupViewModel.kt
@@ -7,9 +7,7 @@ import androidx.lifecycle.viewModelScope
 import com.depromeet.threedays.core.BaseViewModel
 import com.depromeet.threedays.domain.entity.auth.SignupMember
 import com.depromeet.threedays.domain.entity.member.AuthenticationProvider
-import com.depromeet.threedays.domain.repository.NotificationTokenRepository
 import com.depromeet.threedays.domain.usecase.auth.CreateMemberUserCase
-import com.depromeet.threedays.domain.usecase.notification.token.UpdateNotificationTokenUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
@@ -21,8 +19,6 @@ import javax.inject.Inject
 @HiltViewModel
 class SignupViewModel @Inject constructor(
     private val createMemberUserCase: CreateMemberUserCase,
-    private val notificationTokenRepository: NotificationTokenRepository,
-    private val updateNotificationTokenUseCase: UpdateNotificationTokenUseCase,
 ) : BaseViewModel() {
 
    private val _action = MutableSharedFlow<Action>()
@@ -48,17 +44,6 @@ class SignupViewModel @Inject constructor(
             }.onFailure { throwable ->
                 Timber.e("--- SignupViewModel - Signup error: ${throwable.message}")
             }
-        }
-    }
-
-    fun updateFcmToken() {
-        viewModelScope.launch {
-            notificationTokenRepository.getToken()
-                ?.runCatching {
-                    updateNotificationTokenUseCase(notificationToken = this)
-                }?.onFailure { e ->
-                    Timber.e(e, "Failed to update registration token")
-                }
         }
     }
