+++ presentation/signup/src/main/java/com/depromeet/threedays/signup/complete/SignupCompleteActivity.kt
@@ -1,6 +1,7 @@
 package com.depromeet.threedays.signup.complete
 
 import android.os.Bundle
+import androidx.activity.viewModels
 import com.depromeet.threedays.core.BaseActivity
 import com.depromeet.threedays.core.util.setOnSingleClickListener
 import com.depromeet.threedays.navigator.HomeNavigator
@@ -11,12 +12,14 @@ import javax.inject.Inject
 
 @AndroidEntryPoint
 class SignupCompleteActivity : BaseActivity<ActivitySignupCompleteBinding>(R.layout.activity_signup_complete) {
+    private val viewModel by viewModels<SignupCompleteViewModel>()
 
     @Inject
     lateinit var homeNavigator: HomeNavigator
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        viewModel.updateFcmToken()
         initView()
     }
